fix(esp32-node): TickType::new_millis takes u64, not u32
cargo build --release caught it immediately: esp-idf-hal 0.45.2's
TickType::new_millis(ms: u64) doesn't take the u32 timeout_ms was
computed as. Cast at the one call site.

+++ esp32-node/src/dio0_interrupt.rs
@@ -76,7 +76,7 @@ impl<'d> sx127x::DioWait for Dio0Interrupt<'d> {
         // FreeRTOS tick this ultimately becomes can't represent
         // microsecond precision anyway, so this only ever rounds up to the
         // next whole millisecond, never down.
-        let timeout_ms = timeout_us.div_ceil(1000).max(1);
+        let timeout_ms = timeout_us.div_ceil(1000).max(1) as u64;
         let notified = self.notification.wait(TickType::new_millis(timeout_ms).ticks());
         Ok(notified.is_some())
     }
